+++ client/src/pages/about-us/GettingInvolvedSection.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const GettingInvolvedSection = () => {
+  return (
+    <section
+      id={"getting-involved"}
+      className={"tw-min-h-screen tw-bg-primary-blue"}
+    ></section>
+  );
+};
+
+export default GettingInvolvedSection;
