+++ client/src/pages/about-us/page.js
@@ -1,0 +1,20 @@
+import React from "react";
+import LandingSection from "./LandingSection";
+import OurMissionSection from "./OurMissionSection";
+import InvestigatorsSection from "./InvestigatorsSection";
+import StudentTeamSection from "./StudentTeamSection";
+import GettingInvolved from "src/components/all-components/GettingInvolved";
+
+const AboutUs = () => {
+  return (
+    <div>
+      <LandingSection />
+      <OurMissionSection />
+      <InvestigatorsSection />
+      <StudentTeamSection />
+      <GettingInvolved />
+    </div>
+  );
+};
+
+export default AboutUs;
